Simplify analogy
The comparison to UDP sockets is not as good as a comparison to streams.
UDP sockets are sometimes spun up in the fashion described, but they are
also often a fixed thing, such as the socket that a DNS server listens
on.  That evokes the wrong model in which the caller chooses and there
is some possibility of persistence.