Two layer design with CAPSULE
This PR aims to take the design discussions we had at the last
interim, and propose a potential solution. This PR should in theory
contain enough information to allow folks to implement this new
design. That will allow us to perform interop testing and confirm
whether the design is sound.