Make contexts optional
This change allows clients to disable contexts for a given stream.
This uses a new REGISTER_DATAGRAM_NO_CONTEXT capsule for negotiation.
The important property here is that it requires capsule support even
when contexts are not required. This is critical to the health of the
ecosystem as it will ensure that intermediaries that support datagrams
but not capsules cannot be deployed, as that would prevent future
extensibility.