Rework how datagram format extensibility works
After spending some time on CONNECT-IP, I realized that the
current design won't work well. The primary motivation for
including contexts and formatting in this document (as opposed
to in the individual drafts) is that it allows to build a layer
that can handle extensions without round trip negotiations: in
particular it means clients can start using an extension format
immediately without waiting for the HTTP response. For that to
work, the server needs to drop unknown formats, so we need a
single format enum as opposed to a list of them - that way the
entire context is ignored instead of accepting the context but
ignoring the extension that might have changed the encoding.