Be clearer about message error vs. stream errors
Explicitly replicates parts of H2 and H3 CONNECT stream definition to
make it clear what frame types are/are not allowed on a stream.

Mostly anything else can't happen but if it does, approach error
handling _as if it were like_ a malformed message (I.e whether to
terminate the stream or whatever).